serial port moved into nonVisibleWidgets
