+++ Connection/serialPort.py
@@ -1,106 +1,97 @@
-from DataStructures.makesmithInitFuncs         import   MakesmithInitFuncs
+from kivy.uix.floatlayout                      import  FloatLayout
+from kivy.properties                           import  ListProperty
+from kivy.clock                                import  Clock
+from DataStructures.makesmithInitFuncs         import  MakesmithInitFuncs
+from Connection.serialPortThread               import  SerialPortThread
+
+import sys
 import serial
+import threading
 
-
-class SerialPort(MakesmithInitFuncs):
+class SerialPort(FloatLayout, MakesmithInitFuncs):
+    
+    COMports = ListProperty(("Available Ports:", "None"))
+    
+    def setPort(self, port):
+        print "update ports"
+        print port
+        self.data.comport = port
+    
+    def connect(self, *args):
+        print "connect pressed"
+        
+        self.data.config.set('Makesmith Settings', 'COMport', str(self.data.comport))
+        self.data.config.write()
+        
+        self.recieveMessage()
+    
+    def updatePorts(self, *args):
+        
+        portsList = ["Available Ports:"]
+        
+        for port in self.listSerialPorts():
+            portsList.append(port)
+        
+        if len(portsList) == 1:
+            portsList.append("None")
+        
+        self.COMports = portsList
+    
+    def ports(self):
+        print "ports"
+        self.data.gcode_queue.put("test gcode");
+    
+        '''
+    
+    Serial Connection Functions
+    
     '''
     
-    SerialPort is the thread which handles direct communication with the CNC machine. 
-    SerialPort initializes the connection and then receives
-    and parses messages. These messages are then passed to the main thread via the message_queue 
-    queue where they are added to the GUI
+    def recieveMessage(self):
+        #This function opens the thread which handles the input from the serial port
+        #It only needs to be run once, it is run by connecting to the machine
+        
+        print("Starting Second Thread")
+        #self.data.message_queue is the queue which handles passing CAN messages between threads
+        x = SerialPortThread()
+        x.setUpData(self.data)
+        self.th=threading.Thread(target=x.getmessage)
+        self.th.daemon = True
+        self.th.start()
     
-    '''
+    def listSerialPorts(self):
+        #Detects all the devices connected to the computer. Returns them as an array.
+        import glob
+        if sys.platform.startswith('win'):
+            ports = ['COM' + str(i + 1) for i in range(256)]
 
-        
-    def getmessage (self):
-        print("Waiting for new message")
-        #opens a serial connection called serialCAN
-        from time import sleep
-        
-        try:
-            print("connecting")
-            serialCAN = serial.Serial(self.data.comport, 19200, timeout = .25) #self.data.comport is the com port which is opened
-        except:
-            print(self.data.comport + " is unavailable or in use")
-            self.data.message_queue.put("\n" + self.data.comport + " is unavailable or in use")
+        elif sys.platform.startswith('linux'):
+            # this is to exclude your current terminal "/dev/tty"
+            ports = glob.glob('/dev/tty[A-Za-z]*')
+
+        elif sys.platform.startswith('darwin'):
+            ports = glob.glob('/dev/tty.*')
+
         else:
-            self.data.message_queue.put("\r\nConnected on port " + self.data.comport + "\r\n")
-            gcode = ""
-            msg = ""
-            subReadyFlag = True
-            
-            serialCAN.parity = serial.PARITY_ODD #This is something you have to do to get the connection to open properly. I have no idea why.
-            serialCAN.close()
-            serialCAN.open()
-            serialCAN.close()
-            serialCAN.parity = serial.PARITY_NONE
-            serialCAN.open()
-            
-            print "port open?:"
-            print serialCAN.isOpen()
-            
-            while True:
-                
-                try:
-                    msg = serialCAN.readline()
-                except:
-                    pass
-                try:
-                    msg = msg.decode('utf-8')
-                except:
-                    pass
-                
-                print msg
-                
-                if len(msg) > 0:
-                    
-                    
-                    if msg == "gready\r\n":
-                        subReadyFlag = True
-                        if self.data.gcode_queue.qsize() >= 1:
-                            msg = ""
-                    
-                    if msg == "Clear Buffer\r\n":
-                        print("buffer cleared")
-                        while self.data.gcode_queue.empty() != True:
-                            gcode = self.data.gcode_queue.get_nowait()
-                        gcode = ""
-                        msg = ""
-                    
-                    self.data.message_queue.put(msg)
-                    
-                msg = ""
-                
-                if self.data.gcode_queue.empty() != True and len(gcode) is 0:
-                        gcode = self.data.gcode_queue.get_nowait()
-                        gcode = gcode + " "
-                if self.data.quick_queue.empty() != True:
-                        qcode = self.data.quick_queue.get_nowait()
-                        qcode = qcode.encode()
-                        if qcode == b'Reconnect': #this tells the machine serial thread to close the serial connection
-                            qcode = ""
-                            print("Attempting to Re-establish connection")
-                            serialCAN.close() #closes the serial port
-                            sleep(.25)
-                            try:
-                                serialCAN.open()
-                            except:
-                                return -1
-                        else:
-                            try:
-                                serialCAN.write(qcode)
-                            except:
-                                print("write issue 2")
-                if len(gcode) > 0 and subReadyFlag is True:
-                    gcode = gcode.encode()
-                    print("Sending: ")
-                    print(gcode)
-                    try:
-                        serialCAN.write(gcode)
-                        gcode = ""  
-                    except:
-                        print("write issue")
-                    subReadyFlag = False
-                else:
-                    pass
+            raise EnvironmentError('Unsupported platform')
+
+        result = []
+        for port in ports:
+            try:
+                s = serial.Serial(port)
+                s.close()
+                result.append(port)
+            except (OSError, serial.SerialException):
+                pass
+            except (ValueError):
+                print("Port find error")
+        return result
+    
+    def detectCOMports(self, *args):
+        x = []
+        
+        altPorts = self.listSerialPorts()
+        for z in altPorts:
+            x.append((z,z))
+        
+        self.comPorts = x
