+++ DataStructures/data.py
@@ -24,7 +24,8 @@ class Data(EventDispatcher):
     #The file where logging will take place if it is turned on
     logfile = None
     #This defines which COM port is used
-    comport = "" 
+    comport = ""
+    serialPort = None #this is a pointer to the program serial port object
     #The index of the next unread line of Gcode
     gcodeIndex = 0
     #The amount to move from one step
